rebase (#3)
* Fixed class attribute in markdown (closes #838)

* Fixed plantuml markdown supporto for pelican 3.7 (fixes #833)

* Update Readme.md

Correct minor typo in display math command syntax

* Remove trailing whitespaces from neighbors plugin

* Fixed error: NameError: name 'unicode' is not defined (Python3)

* PEP8 fixes

* Sync just table plugin

* [Fix] Dots after numbers in headers

* Added a small bit of clarification for indexing

* Add github wiki plugin and update readme

* Add author images

* Move bootstrap-rst from submodule to in-tree

I no longer have intent to further this plugin, so I think it's best to move it
in-tree where it's more easily accessible for others.

Since everything is squashed, the changes from the last version are:
  * use relative imports in bootstrap.py (thanks to @davidchall)
  * updated maintainer email
  * very small documentation fixes

* Fix URL

* Update twitter sharing link to 'twitter.com/intent'

* Fix NameError in Python 3

This commit fixes the NameError produced by calling
 filter(lambda string: isinstance(string, basestring))
in Python 3. Because filter() is a generator, the lambda function doesn't
actually called at the point of the filter() statement (which is within
a try/except block), it gets called later and raises the error outside
the try block. With this commit, the code now identifies the relevant
string type (basestring or str) before ever using it, forcing the
NameError to get raised early in a controlled context.

* Parses rdf vocabularies, and processes them as a new media type

This plugin offers a flexible mechanism to write lightweight documentation
about rdf vocabularies, based on sparql queries.

* Add pelican-ert plugin
A plugin for estimating time that could be spent for an article reading.

* Update pelican-toc to 1.2.0

* Update bootstrapify to 1.0.0

* Thumbnailer: add optional input file name regex setting

With the help of the THUMBNAIL_INCLUDE_REGEX setting it is now possible
to restrict thumbnailing to a files matching this regular expression.
The setting is expected to be a string. If available, a regular
expression is created. The default behavior is still that all files not
starting with a dot are respected.

For instance, this can be set to only match files ending in "jpg",
"jpeg" or "png" one can use:

THUMBNAIL_INCLUDE_REGEX = r'(jp[e]g|png)$'

* Interlinks: patch to fix #920 (#925)

* fixes #920

* Interlinks: improves HTML img tag handling. Fixes #886 (again)

* Fix md_inline_extension for Pelican 3.7+ (#875)

* Deprecate feed_summary plugin. Closes #926

* Remove feed_summary plugin from list in ReadMe

* category_meta: Add support for Pelican >= 3.7.0

`category.slug` is not writeable since 648165b839fe9562822fa843cbdbfe224b12dfbd.

* fix a bug with autopages regarding context

* Better tags repartition with TAG_CLOUD_MAX_ITEMS

* Added Support for {attach} images

Handles images with {attach} differently, new feature of Pelican 3.5, current code was not compatible at all as it couldn’t understand where the file was.

* Expand photo size when rotated otherwise the image was being cut. See http://pillow.readthedocs.io/en/3.1.x/reference/Image.html#PIL.Image.Image.rotate for expand setting - this resizes the image when rotating.

It may be applicable to other orientations in this section of code, but I have no photos to test with.

See also https://stackoverflow.com/questions/31303938/how-do-i-stop-pil-from-swapping-height-width-when-rotating-an-image-90

* Fix ASSET_SOURCE_PATHS to be list instead of tuple

Tuple causes error when used, must be list.

* Add --no-headers option

* Handle non-ascii paths in filetime_from_git plugin

Paths with non-ascii characters would give encoding errors.

Fixes #328

* Events: new readme (#948)

Events: New ReadMe file

* Update submodule: pelican-linkclass

* mdx_linkclass.py: Fix typo in comment
* Make the plugin work also for reference-style links
* linkclass.py: Update to Pelican 3.7